Remove VLAN hardcode to-do
Fix the circular import problem preventing us from referencing `DefaultMTLVlan` from the networking package.
